Class splitting test
